Consolidate command logic for aoc_cli
