+++ aoc_cli/__main__.py
@@ -1,3 +1,4 @@
 from aoc_cli.commands.cli import cli
 
-__all__ = ["cli"]
+if __name__ == "__main__":
+    cli()
