+++ aoc_cli/commands/__init__.py
@@ -1,0 +1,7 @@
+from aoc_cli.commands.cli import cli
+from aoc_cli.commands.init import init
+from aoc_cli.commands.uninstall import uninstall
+from aoc_cli.commands.update import update
+from aoc_cli.commands.hosts import update_hosts
+
+__all__ = ["cli", "init", "uninstall", "update", "update_hosts"]
